findmypath-agents/requirements.txt
# LangGraph Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sse-starlette>=1.6.0

# HTTP Client
httpx>=0.26.0

# PDF Generation
reportlab>=4.0.0
python-docx>=1.1.0
jinja2>=3.1.3

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
